+++ go.mod
@@ -8,8 +8,10 @@ require (
 )
 
 require (
-	github.com/andybalholm/brotli v1.0.4 // indirect
-	github.com/klauspost/compress v1.15.0 // indirect
+	github.com/andybalholm/brotli v1.1.0 // indirect
+	github.com/essentialkaos/depsy v1.1.0 // indirect
+	github.com/klauspost/compress v1.17.7 // indirect
 	github.com/valyala/bytebufferpool v1.0.0 // indirect
-	github.com/valyala/fasthttp v1.34.0 // indirect
+	github.com/valyala/fasthttp v1.52.0 // indirect
+	golang.org/x/sys v0.18.0 // indirect
 )
